requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-countries==7.6.1
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
paypalrestsdk==1.13.3
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
Pyng==1.6.1
pyOpenSSL==24.3.0
pyotp==2.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
radis==0.15
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
